feat(Établissement): utilise le nouvel endpoint api pour récupérer les justificatifs de TD (#6983)
